+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## 0.7.6 (ongoing)
 
+- `interp_nearest_xx`、和 `binning2d` 函数移除 `in_channels_last` 和 `out_channels_last` 参数，强制保证 channels first 的输入输出。
+- 修正 `interp_nearest_xx` 函数输入字符串类型 `fill_value` 的报错。
+- 修正 `binning2d` 函数输入多通道 `values` 的报错。
+
 ## 0.7.5 (2025-07-25)
 
 - 用 `frykit.conf` 模块替代 `frykit.option` 模块，通过直接修改 `dataclass` 实例的属性来修改配置，以获得更好的 IDE 提示。例如：